Support for searching Google Chrome history
This uses the WASM compiled version of sqlite since [native modules are
not supported](https://github.com/raycast/extensions/issues/135).

The query for matching is a rudimenary `title LIKE '%query%'` where
clause. I intend to make this a bit more sophisticated to handle
multiple terms.

It is only using the native throttle support to debounce access to
the database. It's possible that this will need more tuning or better
resource management.

It also does not yet support choosing a specific profile for history.
This would be possible by setting that string in preferences. For now it
defaults to "Default".