Improve database lifecycle management
This improves the database lifecycle management. Now, the db is loaded
once when the component is mounted and disposed when the component is
unmounted.

Note that I have not been able to see the disposal logic get called. I
suspect that this is either due to the way that Raycast handles
extension component lifecycles or the dev environment operating differently.