Give project name higher precedence than project path (#111)
* Give project name higher precedence than project path

* Use threshold value for filtering with scoreFn

Typechecker didn't catch this locally, but: the default score function returns
null if there's no match, so I tried to do that too. However, the type requires
that it return a number always. I've replaced null-on-no-match with
below-a-threshold-on-no-match, which appears to work just the same.

I picked the threshold pretty arbitrarily, but if scores are in the (negative)
tens or thousands at most, negative a million is a pretty bad match.