Add more sensible way to deal with encrypted notes
Sergey managed to get the search for encrypted notes working in 376ad6de , however this leads to problems since some of the actions rely on having access to the text of the note and some functions of the official Bear URL scheme are not available for encrypted notes.

I removed actions (which won't work with encrypted notes) from the actions panel if an encrypted note is selected in Raycast.