Use pre-generated JSON
This removes the need to package a binary with the extension, and also improves performance.

One downside here (and this was the case with the binary) is that by checking in `symbols.json`, it needs to be manualy built in order to update the extension.

It would be nice if we could just update the symbol source files and `symbols.json` would be generated on the fly when being built and published.