Dash: Support spaces in query
Fixes #976

Since we weren't quoting the `query` variable, a space in the query
resulted in bash thinking we started a second argument to the
`./dashAlferdWorkflow` call.

At first I opted to just wrap the call in quotes, but that would still
break if a user (for some reason) decided to add a quote to their query
string, so instead I went with `execFile` instead, which takes care of
the quoting/escaping for us. This way we also prevent shell injections
(even though it's your own machine and you're free to do with as you
please), I just figured it was cleaner that way.