Add gif-search extension
This extension allows users to search for animated GIFs via giphy.com or tenor.com. It requires the user to get a (free) API token for each service they intend on using. Once that's set, they can view the top 10 trending gifs, or search for specific ones.

I really, REALLY wanted to have an embedded full-size GIF preview to be a part of this extension. However, Raycast's `<Detail>` component's markdown renderer only displays the first frame.

I tried implementing a custom GIF decoder and renderer in react, with limited amounts of success. It somewhat works, but also has some gnarly bugs. I've decided to hide it behind a setting (disabled by default) in order to be able to submit this extension to the store.

With the preview disabled, all this extension will allow you to do is get a list of gifs, and either open it in your browser, or copy it to your clipboard. Still useful, but not as awesome as if the inline gif renderer worked.