Add "IATA Code Decoder" extension
This commit introduces a new Raycast extension, IATA Code Decoder,
for decoding [IATA][1] codes for airports, airlines and aircraft.

The extension contains a command for airports, a command for
airlines and a command for aircraft.

When you enter an IATA code, it'll tell you what that code refers
to (for example, the airport code `SFO` refers to San Francisco
International Airport). It includes typeahead functionality
to give helpful suggestions - for example, when you enter `77`
as an aircraft code, it'll show you all aircraft with codes
beginning in `77` - the Boeing 77- family.

The data is sourced from a [bespoke API][2].

[1]: https://iata.org
[2]: https://github.com/timrogers/iata-code-decoder-api