Fix typos in README and improve screenshots
Fewer screenshots, and now they're all centered!

Also renames the `openInBrowser` default action. This used to be the default for everyone, and if you used that version of the extension it'll remain the default across updates (even though there's a new default). Renaming it should force the setting to reset so most people get the saner defaults.