Dash: Fast docset filtering from MultiDocsetSearch (#1906)
When the first word of the search string matches one of the docset keywords, we can append a `:` to the keyword and immediately filter the results down to just that docset, similar to the experience with Dash’s Alfred workflow. This solves the speed issue moving from Alfred in my use.