Add extension for controlling the Logitech Litra Glow light
This adds a new Raycast extension for turning on and off the
[Logitech Litra Glow][1] light.

The light connects via USB, and this extension uses [`node-hid`][2]
under the hood to send instructions to the device. `node-hid` is
wrapped by my own [`litra-glow`][3] npm package.

Note that `node-hid` requires some funky binaries for its
operation. With some [great help][4] from @GastroGeek in Slack,
I've managed to get those binaries working with a custom
`prebuild`/`predev` script which moves the files where they need
to be, but I don't know whether this solution will work if this
extension is in the Store, how robust it will be or whether it
will work for people without an M1 Mac.

[1]:
https://www.logitech.com/en-gb/products/lighting/litra-glow.946-000002.html
[2]: https://github.com/node-hid/node-hid
[3]: https://github.com/timrogers/litra-glow
[4]:
https://raycastcommunity.slack.com/archives/C02HEMAF2SJ/p1664118805390089