Changed the SearchList implementation
SearchList currently tries to load all the repositories in a workspace into cache before presenting to the user. This works great for workspaces with small number of repositories but it takes forever for workspaces with larger number of
repositories. Changed the implementation to load the fist page of the data as the user types. swr cache will still cache per query.