[bitbucket] Changed the SearchList implementation to work with workspaces with large number of repositories (#4106)
* Changed the SearchList implementation

SearchList currently tries to load all the repositories in a workspace into cache before presenting to the user. This works great for workspaces with small number of repositories but it takes forever for workspaces with larger number of
repositories. Changed the implementation to load the fist page of the data as the user types. swr cache will still cache per query.

* fixed build

* fixed lint errors

* ran npm audit

* removed default values from preferences

* removed placeholder for password and added link in the description

* Update package.json

* Changelog

Co-authored-by: Per Nielsen Tikær <per@raycast.com>