StoicQuotes: Get a random quote
Get a random quote from one of the stoic philosophers - Marcus
Aurelius, Cato, Zeno, Seneca or Epictetus.

This extension uses the StoicQuotes API: https://stoicquotesapi.com

Each time the command is invoked, a new random quote is displayed.

The extension also offers to copy the quote either as plain text or as
markdown.