Adding support for more than one Key Light
I updated the core parts of the extension to remove the assumption that there is a single Elgato Key Light. To do this I needed to add an array of Key Lights and then modify each function to loop through the array. This also required that the user tell us how many lights they have.