Fix for #4816: add support for SetApp flavor of NotePlan3
NotePlan3 is distributed through AppStore and SetApp. These distribution
methods result in different naming of directories where NotePlan3 stores
notes.
This fix tests for distribution flavor of NotePlan: if SetApp directory
exists then it is used as a location to search for notes, and falls back
to AppStore location otherwise