1Password: make primary and secondary actions configurable
Added two dropdowns to the extension's configuration to allow users to
customize the primary and secondary actions for Login items.

The default behaviour is unchanged: "Open in 1Password" is the default
primary action, and "Open in Browser" is the default secondary action.

Added keyboard shortcuts to the "Open in 1Password" and "Open in
Browser" actions, for use when they are not the primary/secondary
actions.

For "Open in Browser", the existing behaviour is maintained: the action
is only available if the item has a primary URL.

For non-Login items, the behaviour is unchanged: the only available
action is "Open in 1Password".

Note: the new configuration is only applicable when using 1Password v8.

Also:
* refactored the action panel into a separate component
* made the `Item.url` field optional to match the format of the
  1Password CLIs' JSON output