Fix bugs and add color palettes
Fixed the problem of automatic refresh when the dynamically generated swatch is loaded, ensuring that the refresh of the palette can only be triggered manually.
A customizable built-in color palette has been added. Users who need to customize the color palette can easily customize it by modifying the Json file without modifying the code.