feat: use dedicated RecentPage object
originally, this was supposed to be just a fix for the last_edited_time
being shown incorrectly in the page details view. But, it turns out that
it required a major change because before, recent pages were saved by
modifying the page's last_edited_time, which caused there to be a
disparity between the real last_edited_time in Notion, and the locally
saved last_edited_time.

for example, suppose page A was in database D, and page A was last
visited a minute ago, but last edited 2 hours ago. In the search view
(and page details), it would show up as last edited a minute ago
(incorrect) But, in the database view, it would show up last edited two
hours ago (correct), because the last_edited_time is pulled directly
from Notion.

this also fixes the issue where if a page title was changed, the recent
page would still show the old title, and cause duplication (i think).

now, recent pages are saved by saving the page id and the
last_visited_time (and the page type). this is much more accurate and
should fix the issues mentioned above, and other related issues with
storing an entire page object locally.