Update hide-all-apps extension: speed up hiding of other applications (#10551)
* perf: speed up hiding of other applications

* chore: Add missing CHANGELOG.md entry

* refactor: invoke 'Hide Others' via explicit menu command

In a follow-up to improve the performance of hiding other applications,
instead of iterating through open processes and setting their visibility
to false, we're continuing to use the native 'Hide Others' command. But,
instead of using the keyboard shortcut, which was problematic given that
it could be overridden, we're now interacting with the menu command
directly via System Events.