[Toggl Track] Improve `CreateTimeEntryForm` (#11133)
* ⚰️ Remove unused properties from CreateTimeEntryFormParams

* 🐛 Refresh time entries after creating a new one

Starting a new "running time entry" might end the current one. The ended time entry won't show in the list unless the time entries are refreshed.

* 🚸 Organize form fields

* 🚸 Remove icons from tasks

* ♻️ Use uncontrolled field for billable

* ✨ Add fields for client and project instead of using project groups

* ✨ Add workspace field

* 📝  Update changelog

* 🐛 Update client dropdown options on "No Project" selection

* 🔥 Remove accossiate projects from task dropdown options

The tasks field is only shown when a project has tasks. I will likely modify this funtionality when implementing the ability to create a new task from this form.