[ai-git-assistant] Add support for generating PR descriptions  (#11886)
* Add support for generating PR descriptions

This commit adds support for generating PR descriptions using an AI. The extension now has a new feature called "Generate PR Description" which takes the path to the Git repository from the selected directory in the Finder and calls an Raycast AI with key context details about current repository branch changes. It also provides a facility to regenerate the PR description and allows direct pasting of the generated PR description to the active application.

* Update screenshots

* Update changelog

* Fix code style

* Update the extension icon

* Update CHANGELOG.md and optimise images

---------

Co-authored-by: raycastbot <bot@raycast.com>