Add `/status_name` to todo creation parser
I've oftentimes found myself wanting to set a status when creating a TODO, for e.g. if it's something I'm already in the middle of but forgot to add. This PR suggests using a leading `/` to indicate setting a specific status to the newly created item.

I also toyed with using `s:` for the parser, but `/` is one less character so went with that. Am happy to change it to something else if others feel strongly about it.