chore(dust-tt): infer username from email
Previously, all API calls were made with a misleading
context, which made all requests appear as if they
were all made by the same user `Raycast`.

In this commit, we're using the user's email to
_infer_ a username, and use this value when making
API calls.

The immediate benefit of this is that the _usage_
API endpoint [1] will now allow to distinguish between
multiple Raycast users, and will allow to accurately
track usage and adoption.

[1]: https://docs.dust.tt/reference/get_api-v1-w-wid-workspace-usage