Add Elixir extension (#15580)
* feat: add elixir extension

* fix(lint): add eslint configuration and fix lint errors

* refactor: improve liste item loading to prevent flickering

* Update the way we're loading list items for both the
`search-functions` and `search-modules` commands, as well as the
`ModuleDetail` component, so as to leverage promises and use the
`<List>` component `isLoading` prop, that way we can avoid the empty
state flickering.

* chore: remove empty component file

* Delete `src/components/ListItem.tsx` as this file was completely
empty. This was probably committed on accident when creating the
`ListItems` component, which itself has a `ListItem` function
responsible for creating the element for a list item.

* Update CHANGELOG.md and optimise images

---------

Co-authored-by: raycastbot <bot@raycast.com>