Remove superfluous app check. (#16283)
* Remove superfluous app check.

This app check is hanging for some reason, even though the osascript
command works without issue. It is nevertheless, completely unnecessary
and I suspect just a product of copypasta from similar extensions.
Furthermore, it was only used in the new tab function, and not similarly
in the history tab function, so it just seems like cruft.

Fixes #16280.

* Update CHANGELOG.md

* Update CHANGELOG.md and optimise images

---------

Co-authored-by: Per Nielsen Tikær <per@raycast.com>
Co-authored-by: raycastbot <bot@raycast.com>