Miniflux: Add mark as read actions (#17038)
* Copy updates.

I noticed that there were a number of small
improvements that could be made when first using
the extension; typos, consistency of naming,
removal of "!".

* Ensure correct toast style is being used.

* Add contributors to extension.

* Remove unused code.

* Use PreferenceValues.

The existing code had a number of places where
the Preferences type (that does not exist) was
being used when calling getPreferenceValues().
Instead we should use the PreferenceValues type.

* Avoid using hook conditionally.

Within the ReadRecentEntries command, where there
was no cached content yet (on first load), the
useEntryIcon hook was being called conditionally,
which meant that when the feed entries were
fetched, they then were not being rendered. You
would need to reopen the command to see them.

* Refactor updateEntries to updateEntry.

This function was only able to update a single
entry at a time, with a given entry id. So I
renamed it to updateEntry, and added a new
function updateEntries that updates multiple
entries at once.

* Add mark entry as read action.

This works in a similar way to the existing
bookmark action, but will instead just mark the
current highlighted entry as read, and then reload
the list of entries in the view.

* Add similar command to mark page as read.

Similar to the previous mark entry as read action,
this will instead mark all entries visible in the
current list view as read, and then reload the
view.

* Update changelog.

* Use PR_MERGE_DATE.

* Addressing comments in "Open Miniflux" command

- Remove `null` return value as not needed.
- Remove `await` as not needed.

* Define MinifluxApiError as child class of Error.

* Update extensions/miniflux/CHANGELOG.md

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>

* Pass in correct params to handle functions.

* Show toast if no base URL is set when opening.

* Update CHANGELOG.md and optimise images

---------

Co-authored-by: Per Nielsen Tikær <per@raycast.com>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>
Co-authored-by: raycastbot <bot@raycast.com>