Add new daily note preference to open from a specified vault if given.
Rather than moving the daily note vault preference to the top level and
sharing, this adds it to the daily note command again. This way, it can
be set explicitly for the append command while still allowing this
command to open the selector.

Nevertheless, this gives users the option of setting the vault in which
daily notes should be opened. For my workflow, though I have multiple
vaults, I always use the daily from one, so having to select the vault
when I use the hotkey is an extra step. With this, users can continue to
use the extensions exactly as before, but others can choose to specify
which vault to use for dailies.