Now Playing View - YoutubeMusic only
This is a simple view of the now playing song.
Currently supports only for YoutubeMusic.

This view allows:
- Nice UI for the now playing song with song details.
- Play/Pause button.
- Fast forward/Rewind buttons.
- Next/Previous track buttons.
- Like/Unlike buttons.
- Dislike/Undislike buttons.
- Mute/Unmute button.
- Volume Up/Down buttons.

I tried to reuse the exiting actions as much as possible. but sometimes I had to
create new actions to allow undo options.

Arguments added to preferences:
- Now playing render interval.

I also add the ability to open YoutubeMusic or Youtube in case
no tab is found.