Add action to generate and copy private links
Adds a new command action in "Share Meeting Links" view that will
generate a new one-time use private link and copy it to clipboard. The
action can also be triggered via Cmd + s shortcut. The request can be
cancelled manually or if the same action is triggered on another
meeting. Feedback is shown during the generation process via toasts/HUD.

This API seems to be only available in V2, so new Axios client + helpers
are added to support the feature.

Other changes:
- Ignore raycast-env.d.ts as per Raycast docs
- Updated package.json to add contributor
- Fix lint error and add publish script